# ts-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fulcrumaxe ts-backend API",
    "version": "0.1.0",
    "description": "TypeScript/Bun+Hono backend — mirrors the Python backend with 1:1 parity. Runs on 127.0.0.1:19099 (loopback-only). Auth: set AF_API_AUTH_KEY to enable; unset = auth disabled. SSE routes (/feed, /events) accept ?token= for EventSource clients that cannot send custom headers. See https://github.com/fulcrumaxe/fulcrumaxe/discussions/1437 for the full phased migration plan.",
    "contact": {
      "url": "https://github.com/fulcrumaxe/fulcrumaxe"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:19099",
      "description": "Default loopback server (TS_BACKEND_PORT overrides 19099)"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/graphql": {
      "post": {
        "summary": "Home-grown GraphQL",
        "description": "Hand-rolled GraphQL endpoint with 1:1 parity to the Python graphql_api.py implementation. Supported root query fields: health, budget, cost, registry, agents, kpi, control, audit, replays, spawnQueue, notifications, plugins. __schema and __type introspection are supported. Auth: bearer token + RBAC. Note: several fields are always null due to documented key-name mismatches in the Python reference (budget.used, cost.total_usd, registry.stats.open, etc.).",
        "operationId": "graphql",
        "tags": [
          "GraphQL"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphqlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GraphQL response. data may coexist with errors when partial results are available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphqlResponse"
                }
              }
            }
          },
          "400": {
            "description": "'query' field missing from request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "401": {
            "description": "No credentials provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid token or RBAC denial",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from AF_API_AUTH_KEY. Required when auth is enabled. SSE routes (/feed, /events) also accept ?token=<key> as a query param because EventSource clients cannot set custom headers."
      }
    },
    "schemas": {
      "DetailError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "GraphqlRequest": {
        "type": "object",
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "GraphQL query string. Supported root fields: health, budget, cost, registry, agents, kpi, control, audit, replays, spawnQueue, notifications, plugins. __schema and __type introspection are supported.",
            "example": "{ health { ok loop { healthy } } }"
          }
        }
      },
      "GraphqlResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Present when the query executed (may coexist with errors)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                }
              }
            },
            "description": "Present when one or more field errors occurred"
          }
        }
      },
...
    }
  }
}